Add support for the reverse operation to the ntcsim component
With this change the ntcsim component can calculate the resistance
from a temperature required to make NTC's "smart".

This component get's a new 'output' property to which the normalized
resistance value (0..1.0) will be send to and the property
'output_resistance' to specify the nominal resistance value of the
simulated NTC. The coefficients for the NTC are specified by the
'calibration' property.